Introduced new package for logger
