+++ logger/std_logger.go
@@ -1,4 +1,4 @@
-package raft
+package logger
 
 import (
 	"fmt"
